Fix for edit module inconsistency with other edit commands
